+++ extensions/data-plane-transfer/data-plane-transfer-sync/src/test/java/org/eclipse/dataspaceconnector/transfer/dataplane/sync/api/rules/ContractValidationRuleTest.java
@@ -1,0 +1,104 @@
+/*
+ *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.transfer.dataplane.sync.api.rules;
+
+import com.nimbusds.jose.JWSAlgorithm;
+import com.nimbusds.jose.JWSHeader;
+import com.nimbusds.jwt.JWTClaimsSet;
+import com.nimbusds.jwt.SignedJWT;
+import org.eclipse.dataspaceconnector.spi.contract.negotiation.store.ContractNegotiationStore;
+import org.eclipse.dataspaceconnector.spi.types.domain.contract.agreement.ContractAgreement;
+import org.jetbrains.annotations.NotNull;
+import org.junit.jupiter.api.Test;
+
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
+import java.util.UUID;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.eclipse.dataspaceconnector.dataplane.spi.DataPlaneConstants.CONTRACT_ID;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+class ContractValidationRuleTest {
+
+    private final ContractNegotiationStore contractNegotiationStore = mock(ContractNegotiationStore.class);
+    private final ContractValidationRule rule = new ContractValidationRule(contractNegotiationStore);
+
+    @Test
+    void shouldSucceedIfContractIsStillValid() {
+        var contractAgreement = createContractAgreement("contractId", Instant.now().plus(1, ChronoUnit.SECONDS));
+        when(contractNegotiationStore.findContractAgreement("contractId")).thenReturn(contractAgreement);
+        var claims = new JWTClaimsSet.Builder()
+                .claim(CONTRACT_ID, "contractId")
+                .build();
+
+        var result = rule.checkRule(createJwtWith(claims));
+
+        assertThat(result.succeeded()).isTrue();
+    }
+
+    @Test
+    void shouldFailIfContractIsExpired() {
+        var contractAgreement = createContractAgreement("contractId", Instant.now().minus(1, ChronoUnit.SECONDS));
+        when(contractNegotiationStore.findContractAgreement("contractId")).thenReturn(contractAgreement);
+        var claims = new JWTClaimsSet.Builder()
+                .claim(CONTRACT_ID, "contractId")
+                .build();
+
+        var result = rule.checkRule(createJwtWith(claims));
+
+        assertThat(result.failed()).isTrue();
+    }
+
+    @Test
+    void shouldFailIfContractIdClaimIsMissing() {
+        var claims = new JWTClaimsSet.Builder().build();
+
+        var result = rule.checkRule(createJwtWith(claims));
+
+        assertThat(result.succeeded()).isFalse();
+    }
+
+    @Test
+    void shouldFailIfContractIdContractDoesNotExist() {
+        when(contractNegotiationStore.findContractAgreement(any())).thenReturn(null);
+        var claims = new JWTClaimsSet.Builder()
+                .claim(CONTRACT_ID, "unexistentContractId")
+                .build();
+
+        var result = rule.checkRule(createJwtWith(claims));
+
+        assertThat(result.succeeded()).isFalse();
+    }
+
+    @NotNull
+    private SignedJWT createJwtWith(JWTClaimsSet claims) {
+        var jwsHeader = new JWSHeader.Builder(JWSAlgorithm.RS256).build();
+        return new SignedJWT(jwsHeader, claims);
+    }
+
+    private ContractAgreement createContractAgreement(String contractId, Instant endDate) {
+        return ContractAgreement.Builder.newInstance()
+                .id(contractId)
+                .assetId(UUID.randomUUID().toString())
+                .policyId(UUID.randomUUID().toString())
+                .contractEndDate(endDate.getEpochSecond())
+                .consumerAgentId("any")
+                .providerAgentId("any")
+                .build();
+    }
+}