+++ CHANGELOG.md
@@ -26,6 +26,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Dependency injection using factory/provider methods (#1056)
 * Provisioned resource information in Data Management API (#1221)
 * Add custom Jackson (de)serializer for `XMLGregorianCalendar` (#1226)
+* Add contract validation rule (#1239)
 
 #### Changed